backend/requirements.txt
Django 
djangorestframework 
djangorestframework-simplejwt 
django-cors-headers 
mysqlclient 
stripe 
requests 
Pillow 
PyJWT 
asgiref 
sqlparse
tzdata 
typing_extensions 
urllib3 
certifi 
charset-normalizer
idna 
pexpect
ptyprocess
